+++ src/services/index.ts
@@ -1,5 +1,6 @@
 import type { AxiosRequestConfig } from "axios";
 import axios from "axios";
+import { cookieService } from "@/utils/cookieService";
 
 // TODO: 환경변수 처리
 const BASE_URL = "https://dev.steadies.kr";
@@ -14,30 +15,3 @@ axiosInstance.interceptors.request.use(async (config) => {
   }
   return config;
 });
-
-export const cookieService = {
-  get: async (key: string) => {
-    if (typeof window === "undefined") {
-      const { cookies } = await import("next/headers");
-      return cookies().get(key);
-    }
-    const { getCookie: get_cookie } = await import("cookies-next");
-    return get_cookie(key);
-  },
-  set: async (key: string, value: string) => {
-    if (typeof window === "undefined") {
-      const { cookies } = await import("next/headers");
-      return cookies().set(key, value);
-    }
-    const { setCookie: set_cookie } = await import("cookies-next");
-    return set_cookie(key, value);
-  },
-  delete: async (key: string) => {
-    if (typeof window === "undefined") {
-      const { cookies } = await import("next/headers");
-      return cookies().delete(key);
-    }
-    const { deleteCookie: delete_cookie } = await import("cookies-next");
-    return delete_cookie(key);
-  },
-};
